FEATURE<Server>: implement Portfolio Domain
Merge: Merge branch 'BE/feature/portfolio' into 'develop'

* FEAT: Implement PortfolioInformationViewUseCase

* FEAT: Implement PortfolioBundle For PortfolioInformation

* FEAT: Implement MemberBundle For PortfolioInformation

* REFACTOR: Add From Method

* REFACTOR: refactor EachRank

* REFACTOR: Change Rank Data Type (Long To Double)

* FEAT: Implement PortfolioInformation API

* FEAT: Implement PortfolioInformation API

* FEAT: Implement CorporationDetailsProvider

* FEAT: Add PortfolioRepository Method

* FEAT: Add PortfolioRepository Method

* FEAT: Implement PortfolioBundle Domain logic

* FEAT: Implement AssetsBundle Domain logic

* SAVE

* REFACTOR: change Url Pattern

* REFACTOR: Add PortFolioModifyApiRequestDto Validation

* REFACTOR: Change JPQL To Data JPA Method

* REFACTOR: Change SnakeCase To CamelCase

* FEAT: add Portfolio Entity

* FEAT: implement portfolioApi

* FEAT: implement portfolioException

* FEAT: implement portfolioInformation

* FEAT: implement PortfolioRepository

* FEAT: implement portfolioRemove

* FEAT: implement portfolioModify

* FEAT: implement portfolioCreate

* FEAT: implement portfolioAdd


See merge request s09-fintech-finance-sub2/S09P22A309!23

+++ backEnd/api-server/src/main/java/hasix/junear/corporation/application/dto/ViewCorporationDetailsResponse.java
@@ -11,18 +11,18 @@ public class ViewCorporationDetailsResponse {
     private String industryType;
     private String corporationCode;
     private String name;
-    private Long stabilityRank;
-    private Long growthRank;
-    private Long profitabilityRank;
-    private Long activityRank;
+    private Double stabilityRank;
+    private Double growthRank;
+    private Double profitabilityRank;
+    private Double activityRank;
     private Double totalRank;
     private Long stockClose;
 
 
     @Builder
     public ViewCorporationDetailsResponse(Long corporationId, Long industryId, String industryType, String corporationCode,
-            String name, Long stabilityRank, Long growthRank, Long profitabilityRank,
-            Long activityRank,
+            String name, Double stabilityRank, Double growthRank, Double profitabilityRank,
+            Double activityRank,
             Double totalRank, Long stockClose) {
         this.corporationId = corporationId;
         this.industryId = industryId;
@@ -36,19 +36,4 @@ public class ViewCorporationDetailsResponse {
         this.totalRank = totalRank;
         this.stockClose = stockClose;
     }
-
-//    public static ViewCorporationDetailsResponse from(Corporation corporation) {
-//        return ViewCorporationDetailsResponse.builder()
-//                .corporationId(corporation.getId())
-//                .industryType(corporation.getIndustryId())
-//                .corporationCode(corporation.getCorporationCode())
-//                .name(corporation.getName())
-//                .stabilityRank(corporation.getStabilityRank())
-//                .growthRank(corporation.getGrowthRank())
-//                .profitabilityRank(corporation.getProfitabilityRank())
-//                .activityRank(corporation.getActivityRank())
-//                .totalRank(corporation.getTotalRank())
-//                .stockClose(corporation.getStockClose())
-//                .build();
-//    }
 }
